requirements.txt
asgiref==3.8.1
Django==5.1.1
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-seed==0.3.1
djangorestframework==3.15.2
Faker==28.4.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
toposort==1.10
